requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
greenlet==3.0.3
gunicorn==21.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.23.0
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.4
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rapidfuzz==3.6.1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzdata==2023.3
Werkzeug==3.0.1
zipp==3.17.0
